- Cache Handling:
The existing caching approach with event.attributes has been maintained as it was previously working without issues.

-Pagination Logic:

 The loop in get_paginated_tasks now breaks if the response.get("last_page", False) condition is met, indicating the last page.

 Webhook Logic:

 - Webhook resolution logic has been removed from the client class, while the get_webhooks and create_webhook methods handle only the fetching and creation of webhooks.
